refactor: extract common parts to ess/reflectometry
